feat(broadcast): initial implementation of a periodic cleanup job
We're not notified when the user deletes a shortcut (rendering the backing shortcut DB object useless), but we can poll this information. This implements a simple periodic daily* job to do exactly that: clean out stale DB entries that aren't referenced anymore.
WARNING: largely untested! Mostly committing for myself. Before merging I want to have found a way to force-trigger a work job; perhaps this consists of a WorkerTestActivity scheduling the given work class as a oneshot, that is only compiled in development builds (if that is possible)