fix(broadcast): cleanup job works now
Apparently there's a weird interaction between Hilt injection for workers (requiring a custom WorkManager Configuration) and something called "App Startup"; the latter needs to be disabled (or the WorkManager portion removed) for our config change to work. And for THAT to work, apparently another Hilt Compiler dependency is required (for whatever reason). I know this sounds absurd, but HiltWorkers should just work now.

Additionally, I went through and made a (generic) debug activity that simply one-shot executes a work it is defined for (in the DEV manifest, which is only merged in dev builds). It can be executed, for example, using ActivityManager itself!